Kucoin/requirements.txt
kucoin-python==1.0.11
numpy==1.24.1
websocket-client==1.4.2
requests==2.28.1
rel==0.4.7
logtail-python==0.1.3
tenacity==8.1.0
